Move StarIcon out of WordCard
